requirements.txt
# Base dependencies
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0
pycocotools==2.0.8

# Demo dependencies
matplotlib>=3.9.1
jupyter>=1.0.0
opencv-python>=4.7.0

# Dev dependencies
black==24.2.0
usort==1.0.2
ufmt==2.0.0b2

# Additional dependencies
scikit-image>=0.23.2
scikit-learn>=1.5.1
pandas>=2.2.2
scipy>=1.14.1
jsonlines>=3.1.0
albucore==0.0.24
albumentations==2.0.8
nibabel==5.2.1
